Fix implicit 64-to-32 bit integer narrowing
Summary:
These fire when a 64-bit integer (e.g., `size_t`, `int64_t`) is implicitly converted to a 32-bit type (e.g., `int`, `uint32_t`), which can cause data loss.

Fixes applied:
- Widened variable types to match source (e.g., `int` → `size_t`)
- Added explicit `static_cast` where narrowing is intentional and safe

Differential Revision: D97079855

+++ momentum/io/marker/c3d_io.cpp
@@ -138,7 +138,7 @@ loadC3d(const std::string& filename, UpVector up, std::span<const std::string> v
 
     // Go through the labels to find all the subjects and their markers.
     // A Subject is a collection of points that should be grouped together to represent an object.
-    std::unordered_map<std::string, std::vector<int>> subjectNameMap;
+    std::unordered_map<std::string, std::vector<size_t>> subjectNameMap;
     subjectNameMap.reserve(kPointsPerFrame);
     for (size_t iLabel = 0; iLabel < kPointsPerFrame; ++iLabel) {
       const std::string& name = pointLabels[iLabel];
